requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1.1
djangorestframework==3.11.1
pytz==2020.1
wincertstore==0.2
jsonpickle==1.4.1
